refactor(tools.general): change name from general_tools
